+++ library/test/org/webpki/sks/ws/client/SKSWSClient.java
@@ -216,7 +216,7 @@ public class SKSWSClient implements SecureKeyStore, WSSpecific {
                                                          String issuer_uri,
                                                          PublicKey keyManagementKey,
                                                          int clientTime,
-                                                         int sessionLifeTime,
+                                                         short sessionLifeTime,
                                                          short sessionKeyLimit,
                                                          byte[] serverCertificate) throws SKSException {
         try {
